+++ blog/main/templates/main/books.html
@@ -17,6 +17,7 @@
                     <th scope="col">ID Книги</th>
                     <th scope="col">Название</th>
                     <th scope="col">Имя автора</th>
+                    <th scope="col">Категория</th>
                 </tr>
                 </thead>
                 <tbody>
@@ -25,6 +26,7 @@
                     <td>{{ book.id }}</td>
                     <td>{{ book.title }}</td>
                     <td>[{{ book.author.name }}]</td>
+                    <td>--{{ book.category.name }}--</td>
                 </tr>
                 {% endfor %}
                 </tbody>
